Moved Actions and waitlist to components folder.
