+++ FrontEnd/src/Components/Clase/Actions.jsx
