+++ FrontEnd/src/Components/Clase/WaitList.jsx
